+++ dotnet/Data/IAvailabilityRepository.cs
@@ -15,5 +15,6 @@ namespace AvailabilityNotify.Services
         Task<bool> SaveNotifyRequest(NotifyRequest notifyRequest, RequestContext requestContext);
         Task<NotifyRequest[]> ListRequestsForSkuId(string skuId, RequestContext requestContext);
         Task<NotifyRequest[]> ListNotifyRequests();
+        Task<NotifyRequest[]> ListUnsentNotifyRequests();
     }
 }