+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Processing routes
+
 ## [0.0.8] - 2021-06-08
 
 ## [0.0.7] - 2021-06-07